extract got code into it's own file
